refactor: transfer example
